# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MCP Tool Evaluation System API
  description: |
    REST API for evaluating Model Context Protocol (MCP) tool effectiveness.
    
    **Features**:
    - Tool selection evaluation with LLM providers
    - Parameter validation (completeness, correctness, type conformance)
    - Inter-tool similarity detection and recommendations
    - Comprehensive metrics (precision, recall, F1, parameter accuracy)
    - Security-first architecture (runtime API keys, never persisted)
    - Fresh tool ingestion per test run for accuracy
    
    **Generated**: 2025-11-10
    **Git Commit**: 859d728254fc64586c05039908056c6b4bec1709
  version: 2.0.0
  contact:
    name: mcp-tef Team
    url: https://github.com/StacklokLabs/mcp-tef
servers:
  - url: https://localhost:8000
    description: Local development server (HTTPS with auto-generated cert)
  - url: http://localhost:8000
    description: Local development server (HTTP, use --no-tls-enabled)
...
paths:
...
  /tools:
    post:
      tags: [tools]
      summary: Create tool definition
      operationId: createTool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolDefinitionCreate'
      responses:
        '201':
          description: Tool created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolDefinitionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'

    get:
      tags: [tools]
      summary: List all tools
      operationId: listTools
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - name: test_run_id
          in: query
          description: Filter by test run ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of tools
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ToolDefinitionResponse'
                  total:
                    type: integer
...
components:
  parameters:
...
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 100
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
  schemas:
...
    ToolDefinitionCreate:
      type: object
      required: [name, description, input_schema, mcp_server_id, test_run_id]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        input_schema:
          type: object
          description: JSON Schema for tool parameters
        output_schema:
          type: object
          nullable: true
        mcp_server_id:
          type: string
          format: uuid
        test_run_id:
          type: string
          format: uuid
          description: '004: Required - links tool to test run that ingested it'
    ToolDefinitionResponse:
      type: object
      required: [id, name, description, input_schema, mcp_server_id, test_run_id, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        input_schema:
          type: object
        output_schema:
          type: object
          nullable: true
        mcp_server_id:
          type: string
          format: uuid
        test_run_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ==========================================================================
    # TEST CASES (001 base, 003: removed model_id)
    # ==========================================================================
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security: []